% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap_fusika.R
\name{plot_heatmap_fusika}
\alias{plot_heatmap_fusika}
\title{Plot a customized heatmap (FUSIKA style)}
\usage{
plot_heatmap_fusika(data, group)
}
\arguments{
\item{data}{A numeric matrix (samples x genes).}

\item{group}{A vector of group labels (length = nrow(data)).}
}
\value{
A ComplexHeatmap object.
}
\description{
Plot a customized heatmap (FUSIKA style)
}
\examples{
set.seed(123)
mat <- matrix(rnorm(40*20), nrow = 40, ncol = 20)
group <- rep(c("Group1", "Group2"), each = 20)
plot_heatmap_fusika(mat, group)

}
