% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_violin_fusika.R
\name{plot_violin_fusika}
\alias{plot_violin_fusika}
\title{Plot a customized violin plot (FUSIKA style)}
\usage{
plot_violin_fusika(
  data,
  group_col,
  value_col,
  title = "Gene",
  show_signif = TRUE,
  signif_offset = 1,
  y_manual = NULL,
  tip_length = 0.01
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{group_col}{A string, name of the column with group labels (x-axis).}

\item{value_col}{A string, name of the column with numeric values (y-axis).}

\item{title}{A string, plot title (usually gene name).}

\item{show_signif}{Logical, whether to show statistical comparisons (default = TRUE).}

\item{signif_offset}{Numeric, how much to offset significance labels above max (used if y_manual is NULL).}

\item{y_manual}{Optional vector of y-positions for significance labels (same length as number of comparisons).}

\item{tip_length}{Numeric, length of bracket tip (default = 0.01)}
}
\value{
A ggplot object of the violin plot.
}
\description{
Plot a customized violin plot (FUSIKA style)
}
